:hammer: Refactor and add RuleBook
